requirements.txt
agno
amadeus 
groq
python-dotenv
setuptools
duckduckgo-search
fastapi
yfinance
uvicorn